+++ backend/src/main/java/com/debede/backend/repository/ValoracionRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ValoracionRepository extends JpaRepository<Valoracion, Long> {
+public interface ValoracionRepository extends JpaRepository<Valoracion, Integer> {
 
 }
